# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stade de Mbour",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cartes/generate": {
            "post": {
                "tags": [
                    "Access Cards"
                ],
                "summary": "Générer une carte d'accès",
                "operationId": "b38b89d6fffa7b9a5049d0fc516f7cc4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "UUID de l'utilisateur si déjà existant",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "f3f9c8c1-72ab-4a6b-9d74-3f4078ce7e84"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "role_id": {
                                        "description": "UUID du rôle si on crée un nouvel utilisateur",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "a2f9c8c1-72ab-4a6b-9d74-3f4078ce7e99"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+221777000111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carte générée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/cartes/block/{card}": {
            "post": {
                "tags": [
                    "Access Cards"
                ],
                "summary": "Bloquer une carte d'accès",
                "operationId": "dead4ec5334e6632d3dc05375fdf054a",
                "parameters": [
                    {
                        "name": "card",
                        "in": "path",
                        "description": "UUID de la carte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carte bloquée avec succès"
                    }
                }
            }
        },
        "/api/cartes/deactivate/{card}": {
            "post": {
                "tags": [
                    "Access Cards"
                ],
                "summary": "Désactiver une carte d'accès",
                "operationId": "c30551a0eac2d061ac81e1ea4eb54fec",
                "parameters": [
                    {
                        "name": "card",
                        "in": "path",
                        "description": "UUID de la carte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carte désactivée avec succès"
                    }
                }
            }
        },
        "/api/cartes/{id}": {
            "get": {
                "tags": [
                    "Access Cards"
                ],
                "summary": "Détails d'une carte d'accès",
                "operationId": "6e22ba68a0ac8ff976a229801d0142fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la carte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la carte d'accès"
                    },
                    "404": {
                        "description": "Carte non trouvée"
                    }
                }
            }
        },
        "/api/cartes/sell/{id}": {
            "post": {
                "tags": [
                    "Access Cards"
                ],
                "summary": "Vendre une carte d'accès",
                "operationId": "11fd386aaa5260f1667a90a51306f67b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la carte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carte vendue avec succès"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}